+++ examples/Client/Client.hs
@@ -64,7 +64,7 @@ defineByDocumentFile
           }
           ...on Human {
               lTime: lifetime
-             # TODO: Profession should be defined only once! - prof: profession
+              prof: profession
           }
         }
       }
@@ -74,10 +74,16 @@ ioRes :: ByteString -> IO ByteString
 ioRes req = do
   print req
   return
-    "{\"data\":{\"deity\":{ \"fullName\": \"name\" }, \"character\":{ \"__typename\":\"Human\", \"lifetime\": \"Lifetime\", \"profession\": \"Artist\" } ,  \"char2\":{ \"__typename\":\"Human\", \"lTime\": \"time\"}  }}"
+    "{\"data\":{\"deity\":{ \"fullName\": \"name\" }, \"character\":{ \"__typename\":\"Human\", \"lifetime\": \"Lifetime\", \"profession\": \"Artist\" } ,  \"char2\":{ \"__typename\":\"Human\", \"lTime\": \"time\", \"prof\": \"Artist\" }  }}"
 
 fetchHero :: IO (Either String GetHero)
-fetchHero = fetch ioRes GetHeroArgs {god = Just Realm {owner = "Zeus", surface = Just 10}, charID = "Hercules"}
+fetchHero =
+  fetch
+    ioRes
+    GetHeroArgs
+      { god = Just Realm {owner = "Zeus", surface = Just 10, realmRec = Nothing, realmProf = Just Artist}
+      , charID = "Hercules"
+      }
 
 fetUser :: (ByteString -> IO ByteString) -> IO (Either String GetUser)
 fetUser api = fetch api userArgs